organize a bit more, fix packages.
