+++ bs_repoman/log.py
@@ -1,0 +1,3 @@
+import logging
+
+logger = logging.getLogger("bs_repoman")
